Including spin-orbit Hamiltonian and interaction with an external magnetic field; beta version; only static

+++ ga_QCModel.f
@@ -47,7 +47,7 @@ implicit none
 type(R_eigen)            , intent(in) :: GA
 integer                  , intent(in) :: up
 integer                  , intent(in) :: down
-real                     , intent(in) :: dE_ref
+real*8                   , intent(in) :: dE_ref
 real          , optional , intent(in) :: weight
 
 !local variables ...
